tx: remove utxos elements from details after creation
The utxos element is no longer needed once the caller is happy with the tx
created by create_transaction. Remove it to reduce memory usage and
minimise client conversion effort.

Additionally, rename m_tx_details to match the other auth handlers so the
details handling can be shifted to the base class to be shared at some point.