create_tx: remove the transaction_size member from returned data
We already return the vsize and weight which are the only useful values
for fee calculation purposes, and the binary size of the transaction
element can be trivially calculated as the transaction hex length / 2.

Also document that the tx is incomplete until after signing and that the
vsize, weight and calculated fee rate are estimates that factor in any
missing/incomplete data.