create_tx: set the reordered addressees before updating tx details.
The tx updating relies on the ordering of the adressees (for now) which
are re-ordered as part of tx creation. Now that we only update the tx
info once, it becomes clear that this loop was operating on the
unchanged addressees - i.e. it would update the wrong outputs if the
addressees were actually re-ordered, for example in a multi-asset send
where the assets are not listed in order.