bcur: integration into gdk
the commit integrates the BlockchainCommons/bc-ur library into gdk APIs

The APIs are integrated into the auth_handlers framework adding two
new commands bcur_encode and bcur_decode.
The auth_handlers is also equipeed with functions to signal to the
caller that more data is required to complete the decoding operation.

The library is integrated in a opt-out fashion: when ENABLE_BCUR is
TRUE, cmake plugs in the bc-ur library, otherwise the encode and decode
APIs default back to simple stubs that always throw exceptions when
poked.