blind_tx: move the preparatory blinding setup into the blind_tx handler
This makes the blind_tx HWWI handler externally stateless. It can now be
driven by HWW capabilities, for example to disallow signing pre-blinded
inputs for devices that don't support it.

As a side effect, ephemeral key generation and use is made entirely
local to the handler, and the "eph_private_key" data exposure in the
tx/request details can be completely removed.