blind_tx: simplify the final vbf generation logic
Instead of differentiating between input and output abfs/vbfs, use the
number of our inputs as passed to asset_final_vbf to do so. This means
we can simply append the output values to the input ones directly, and
remove all the code for holding them separately and flattening them back
into the required format for the calculation.

Having done that, we can set the abf/vbf into the output json when generated
instead of after the fact. Since we have then removed repeated checks for
applicability to the vbf, inline the calls that check this so the comments
about this logic make sense in the code context.

Finally, rename various variables to reflect their new roles and improve
some comments.