rust: Make sure block notification always happen before tx notification
In case the sync have to be repeated because the tip after and before the sync
are different, it could happen a block notification is sent before a tx
notification, this may cause incoherent state in cache and wrong value in utxos
and balance computation.

Moreover, asking the server tip separatedely from updating the cache, makes the
code more clear.