rust: share electrum client between loop iterations
Instead of recreating the electrum client at every loop iteration, this keep
the same instance, unless the connection drop, in this latter case an attempt
to recreate the client is made.

This offers only a minor optimization but it's propedeutic to the subscribe
model where the client store the subscriptions to scripts.