rust: Subscribe to scripts
The electrum model use scripts subscription to optimize server utilization
and real-time feedbacks.

Until 20 (gap_limit) scripts are found empty, all wallet addresses are
subscribed to the server. The subscription returns a script status hash that
represent the script status (hash of its txs/height) or None if the scripts
has never been seen.

If the status is not none the history for that script is requested like it was
always done before, once the history is applied to the store its script status
is saved, so that history fetching is done only if the server sent script status
is different.

Before we were relying on the server status for last view of the wallet txs,
with this is not anymore possible because the view could be partial. We instead
remove the transactions from the cache if new transactions are in conflict with
those.