cmake: move zlib up
cmake FindOpenSSL module silently seeks zlib in the system libraries,
this causes the following explicit call to FindZLIB to fail if the version
we require does not match with the version installed in the system.

So far we got away with it because our runners do not have zlib-dev
installed, so FindOpenSSL was giving up the search for zlib and carried
on.
With the new manylinux image, the house of cards collapsed, zlib-devel
is installed, FindOpenSSL picks it up and cmake marks zlib has FOUND.
However, as the system zlib-devel does not match with our own version of
zlib cmake hits a version mismatch.

FindZLIB must be called before FindOpenSSL if we want cmake to pick our
own zlib version and not the system one.