cmake: avoid picking sqlite3 from system libs
sqlite being a pkg-config-based library makes it impossible for someone
to pick a version installed in custom folders.
pkg-config cannot be forced to search in custom folders, but rather
suggest hints. If your machine happens to have a version of such library
installed in your system, there is no chance that your version will be
picked.

To avoid this flaw, we add a custom definition of the cmake sqlite3
target.

cmake could potentially do better:
https://gitlab.kitware.com/cmake/cmake/-/issues/23045