cmake: removing python Development component
when searching for Python package, cmake used to look for Development
COMPONENT.
But we actually need it only on APPLE platforms as setuptools needs
explicit instructions to link against libpython, while in linux
setuptools seems to be more clever/sensible.

In linux platform, the extra COMPONENT actually prevents cmake from
working in a virtual environment.

Removing it and refactoring a bit the CMakeLists for swig_python
bindings