hww: add a "supports_external_blinding" capability for e.g swaps
Raise an error if creating an externally-blinded tx is attempted when
the signer doesn't support it (for example, creating a swap transaction
on Ledger, or Jade with an insufficiently new firmware).

This capability currently defaults to true but will be switched to false
in the next release. This is to prevent disruption in downstream apps
close to the current gdk release.