rust: blob server integration
When starting the session with `blob_server_url` or with
`blob_server_onion_url`and `use_tor` true with an https blob server, the
state of the wallet including settings and transaction memos is loaded at
startup from the content retrieved from the server.
Every change to wallet state in the current session of the wallet is sent to
the server encrypted so that next device loading the same wallet will find
the same state.
The state of the wallet is not loaded real-time, two concurrent sessions will
not see values updating.

Co-authored-by: noib3 <riccardo.mazzarini@pm.me>