cmake: refactor of java bindings
Thanks to cmake's UseSwig module we can have cmake generating the java
bindings and compile the greenaddress library with the bindings in it.

This commit moves our cmake infrastructure to use the swig module and
removes tools/swig.sh script.
It also refactors a bit the build script so to minimize the amount of
work done for android targets