tx: remove the requirement for tx and addressee ordering to match
Instead, lookup the addressee or change output from the transaction
using its scriptpubkey.

This removes the rather hairy handling of pre-blinded output positions
and allows us to avoid reordering the callers addresees arbitrarily in
the future (addressees should remain in the order given by the caller
throughout tx construction and completion). It also fits better with PSBT.

Add a number of invariant/sanity checks on the data source vs
transaction that were previously missed.

Because this matches on the scriptpubkey, this adds a new,
requirement that addressees are unique in outgoing txs, which is checked
in a later commit in this patch series.