create_transaction: re-work/refactor tx creation
The intent of these changes is making create_transaction more compatible with
PSBT, in particular to use less internally-generated metadata which the PSBT
format doesn't provide. For example, change to looking up addressee and change
positions in the tx instead of relying on tx creation having stored them or
re-ordered the user-provided data to match the tx. As a side effect the code
is made somewhat simpler.

- Simplify BTC UTXO selection while keeping it backwards-compatible.
- Use wally for asset UTXO selection.
- Don't re-order "addresees" from the caller, as caller provided data
  should not change except to have further metadata added if required.
- Don't create/blind change addresses unless we need change for a given asset.
- Fix returned change addresses not having the correct change amount.
- Remove "change_index". If the tx has change for a given asset, the
  "change_address[asset]" element will be present and its "satoshi" field
  will be non-zero.
- Replace tx-level "send_all" and "is_redeposit" with per-output "is_greedy".
  Redeposits are just greedy sends to a wallet address. Any asset or
  set of assets can be individually all sent using a greedy output.

NOTE: This change temporarily removes the top level "satoshi" field and
disables input randomization to keep the changes as small as possible.