create_transaction: reset fee and change amounts before building the tx
It may happen after repeated calls that a change address is generated,
then subsequently not used (for example, if the amount to send is later
set to greedy). Ensure that the unused change will have satoshi=0 so the
caller knows it is not part of the created transaction (other than not
appearing in "transaction_outputs".

Also, if an error occurs, this ensures we don't return incorrect
amounts.

Finally, this sets the fee element to 0 for partial transactions since
they have no fee component.