tx: ensure that transaction outputs have unique scriptpubkeys/addresses
For JSON (create_tx) based transactions, scriptpubkeys (addresses) must
be unique so that we can correctly map between tx outputs and the
corresponding adressee or change address. Note this doesn't affect
PSBT-based txs (which have an explicit ordering in their outputs), so
this isn't a general limitation just an artifact of how the JSON based
tx representation currently works.