rust: get_receive_address: do not generate addresses beyond the GAP_LIMIT
Before, the next address pointer was incremented at every
get_receive_address call. Meanwhile, every sync loop, it was
reset to the last used address.

With this changes, the address pointer is incremented at every
get_receive_address call, but when it reaches the GAP_LIMIT, it
is set to 0, looping on this set of pointers.

This additional counter of given addresses is reset to 0 when
the last used script is updated (which could happen after
receiving a transaction).