docs: auto-generate JSON examples for each network type
Uses the developer-only localtest environments to set up a wallet,
receive some funds etc, and so generate real API results as returned by
the current version of gdk.

Each example is added under a collapsed, expandable section for the
network and session type, so differences can be seen at a glance.

This should prevent the persistent issue of the examples being out of
date, and also highlights areas where single and multisig sessions
return different results (that need to be fixed).