create_tx: remove old_used_utxos, rename signing_inputs->transaction_inputs
Also, for sweeping, restore the private key before returning the signed
tx to the caller, so the caller can determine which inputs are from a
sweep.

With these changes, all tx creation flows now result in the complete set
of inputs being placed in "used_utxos". This corresponds to the
behaviour of the "transaction_outputs" element, and removes the need for
callers to look outside of these 2 elements for details on the
txin/txouts. It also means that the tx creation result much more closely
approximates the structure of PSBT now.

Since the inputs are collated by create_transaction, there is no longer
any need for get_ga_signing_inputs() to do so, and the inputs for signing
can be used directly in HWWI's "required_data/transaction_inputs"
(renamed from "signing_inputs").

As a result, we simplify the places where we previously had to account for
both old and new used utxos.

Finally, make some minor comment/documentation fixes.