network_parameters: adding a configuration prefix for server uri
as wamp_transport is moving to a multi-service configuration,
network_parameters has to change and allow wamp_transport to be
instantiated with different server configurations.

This commit adds config_prefix to the involved network_parameters APIs,
and to wamp_transport constructor. Whenever wamp_transport needs to
retrieve the server URL, server_prefix is passed to network_parameters
which then uses it as prefix to retrieve the URI from its configuration.