proxy: pass the session mutex when fetching proxy settings
When tor is enabled, the lock must be held to fetch its internal proxy.
Make this explicit and the method non-const to make this clearer to
callers. Additionally, we now take the lock on non-tor code paths to
ensure any introduced locking issues are noticed even when testing
without tor enabled.

Based on a patch by Leonardo Comandini <leo@leocomandini.org>.