get_subaccounts: discover the last empty subaccount
Instead of discovering the next subaccount, discover the last
empty one.If the subaccount was created without being discovered,
it will be discovered here.

If the subaccount has been created, then its transcations
(if any) are being synced in the background. Attempting
discovery is only useful to avoid setting `bip44_discovered`
incorrectly. But that could happen only if we get the account
information before the first sync is completed, which is a short
time window. However since we have no control on the sync, we
need to perform the discovery additional network calls.