tx: only remove fee from tx summary totals if we have a wallet input
The satoshi summary gives the net effect on the wallet, excluding the
fee. However for some tx types where the user provides no wallet inputs,
the user is not paying the fee - the provider of the inputs is.

This was affecting sweep txs now that they use the common signing path.
Update the summary calculation to only exclude the fee if the caller
provided a policy asset input.

Note this is not completely correct if the fee is not fully covered by the
users input(s), or there are outputs that spend the fee whether partially or
completely. However (a) the summary calculation has been incorrect for these
cases since non-wallet inputs were added and (b) addressing these cases may
require including the fee in the summary which cannot happen in the timescale
of the current release.