psbt: sign PSBTs by composing sign/send auth handlers
This removes the duplication of signing logic in the psbt_sign
implementation and the dubious temporary script munging in favour of
passing is_partial to the sign call. It also allows usage of sign_with
with PSBTs with upcoming changes.

This is a partial implementation. Although signing now goes through the HWWI
correctly, the handler currently only provides enough data for the software
signer to sign. A real hardware signer requires the output details, which
requires upcoming work to provide.