rust: compute tx notifications after syncing all subaccounts
If we have a transaction involving multiple subaccounts, it can happen
that when we fetch data for one subaccount the transaction is returned,
while when we fetch data for another subaccount, the transaction is not
returned.
Before this commit, if this happened, the transaction notification
would miss one of the subaccounts.

To avoid this we create the transaction notifications after syncing all
subaccounts, even though this comes with a performance penalty.