psbt: get_details/sign: support BTC and HWW/Limits/2FA
Moves the conversion from PSBT to sign_transaction JSON into the Psbt class.

Updates psbt_get_details to support BTC, and to convert the psbt into tx JSON
as used by blind/sign/send_transaction. This provides information on all
inputs and outputs in the tx, plus the fee/tx hex etc, without losing the
mapping from index to input/output. As callers already need to handle the
tx JSON format, this should be trivial for them to parse rather than using
a bespoke JSON format just for the psbt_get_details call.

We now also automatically add any missing UTXO data to the input psbt.

Since get_details now provides us with tx JSON, for signing we just augment
it with the required data before passing it to the standard signing path.
Since we return the tx JSON in addition to the psbt when signing completes,
there is no need to return the UTXOs element in the final result (or for
callers to call psbt_get_details after signing).

The net effect of this change is to support the existing tx signing
features for PSBTs, most notably:
- Standard Bitcoin PSBTs can now be signed.
- HWW can be used to sign PSBTv0, PSBTv2 and PSET (always v2).
- 2FA authentication for signing is handled correctly.
- Standard Bitcoin sends respect spending limits to avoid 2FA.

Note that the field mapping for Elements/Liquid is like not yet
complete/correct, particularly for swaps; this will be improved in
subsequent changes.