signer: derive and encache parent xpubs as well as children
The parent key is the key up to and including the final hardened child.
By also caching these, when we next derive a new child we have a shorter
derivation path and so less work to do. Additionally, when parents are cached
we can in future avoid round trips to HWW for children that share the
same parent.