auth_handler: transparently optimize get_xpubs requests
Converts client requested full xpub paths into the minimal set of parents to
request from the the HWW. This reduces the frequency of calls and amount of
work required from the HWW when fetching xpubs.

As part of this change, auto_auth_handler gains the general ability to
alter its sub-handlers requests and responses.