psbt: allow psbt_get_details to work on psbts with inputs that cannot be unblinded
The previous implementation ignored these inputs and returned results as though
they weren't present. We now return the input marked as unblindable in
the same way as unblindable utxos are returned.

If the unblindable input is not marked as being skipped for signing,
additionally return this as a transaction level error.