sign_tx: support the server-side input script hack for checking segwit-ness
Multisig partial txs require inputs to be segwit, which is implemented by
setting the tx scriptsig to the redeem script before passing the tx to
the Green backend. Implement this hack for the create_transaction path
by requiring "redeem_script" to be present in the inputs, and
overwriting then restoring the script when signing with the Green backend.
Update the PSBT signing path to set "redeem_script" accordingly.

Also locally check for segwitness in the server signing path since
signing can now be split for PSBTs using "sign_with" as for the standard
sign_transaction flow.