psbt: detect and mark change outputs when signing PSBT/PSET
For singlesig, we use the internal chain as a change indicator
unconditionally. For multisig, for each asset that the wallet
contributed to the transaction, we have two possible cases that result
in an output being marked as change:

1. The asset is sent outside of the wallet and to one or more wallet outputs.
2. There is no external send, but multiple wallet outputs for the asset.

In both cases, the first wallet output is marked as change. There will
only be multiple wallet outputs where (1) the caller has deliberately
constructed an inefficient send to the same subaccount with change, or
(2) the caller is sending the asset to a different subaccount with
change going back to the sending subaccount (cross-subaccount spend).

Note we do not attempt to identify (2) at this stage, this may be
considered for a later non-bugfix release.