rust: sync: handle removal of unconfirmed transactions
When an unconfirmed transaction is removed from the (server)
mempool, the "script status" of the scripts involving such
transaction should change.
For instance if the script only involves txid, it can change from
`h("txid:0:")` to `null`.
Unfortunately this does not always happen.
Thus we don't have a way to realize that we should update our
internal cache and specifically remove such transaction.

To handle this case, we explicitly request each unconfirmed
transaction to the server, to determine if it should be removed.
This comes with a performance penalty, but it seems to be necessary.