gdk_rust: switch back to serde_cbor
we switched to ciborium because serde_cbor it's unmantained.
However, serde_cbor never gave us trouble, while ciborium had the
"invalid type: bytes, expected bytes" bug, which has been attempted to be solved
upstream in https://github.com/enarx/ciborium/pull/97
Apart that we have no input on the MR yet, in trying to use the patched version
we encountered the error:
"invalid type: string, expected map"
which is probably due to the fact the patched version is on master and not just
the bug fix from the tagged version we use.
While this later approach could work, ciborium doesn't look that reliable so we
decided to switch back to serde_cbor for the time being.