bcur: append bip44 path suffix to crypto-account descriptors
crypto-account elements must be bip44 compatible, but don't include the
trailing '/<0;1>/*' derivation path in order to save space.

We add '/0/*' instead of '/<0;1>/*' when formatting the descriptors,
because some wallets are apparently not multi-index compatible.

Also rename formatting functions to allow debugging, and make
sure an empty crypto-account has an empty descriptors array.