tx: fix the dummy fee estimation data added when creating transactions
Correctly set dummy data to be the expected size. This reduces the size
difference between initially created transactions and the final signed
transaction to zero in most cases for low-R signers.

Note that signatures can occasionally be smaller by one or more bytes
due to the DER representation of smaller r,s numbers.