tx: estimate using low-s signatures for non-grinding signers.
When a signer does not produce Low-R signatures (for example, HWW using
Anti-Exfil), the signature produced must still be Low-S, in order to
satisfy standardness rules. Thus we should produce our dummy script
and estimate the fee based on a signature 1 byte shorter than the DER
maximum which is High-R, High-S.

Update the code to do this, and rename the input script functions for
clarity.