script: further improve fee estimation for non-low-r signers
Previously the code assumed the green backend signature had the same
low-R status as the users signer. In fact the backend always produces
low-R, low-S signatures in order to reduce on-chain usage and user fees.

Clean up the code around dummy scripts and witnesses slightly, with the
aim of fully encapsulating it in due course.