tx: maintain our dummy signature data after user signing for multisig
Prior to this change, user signing a tx replaced the scriptsig/witness
elements with the expected values for backend signing: a PUSH(0) for
pre-segwit inputs, and only the user signature in the witness.

This has four main negative implications:
1) The computed tx size and thus fee/fee rate were computed incorrectly
   in the tx returned to the caller after user signing.
2) For segwit inputs we do not give the server the redeem script in the
   witness, which it could use to determine segwitness instead of the
   current hack of replacing the scriptsig with the redeem script and
   then restoring it after signing.
3) The server cannot (in the future) determine if an input has already
   been signed by it, and thus avoid re-signing/requiring 2fa/adjusting
   limits correctly for txs signed in multiple passes.
4) The server cannot determine inputs signed by recovery paths, which
   should not be subject to 2FA checks.

Change this approach to maintain the scriptsig/witness in the tx after user
signing. The server signature remains as a dummy signature until the
server is asked to sign, at which point we internally modify the tx to its
current expectations. This modification will be removed in the future once
the server is updated to handle the full witness.

The dummy witnesses being created were incorrectly ordered in some
cases; these have been fixed. We also no longer create a dummy witness
script but instead use the correct one from tx creation onwards.
Finally, we simplify some of the signature handling code as a
result, in particular we re-use the code to set the scriptsig/witness
during creation and post user-signing.