send tx: Interrogate the tx for signatures before signing/sending
If the tx has server signatures already (or is spending expired CSV or
sweep outputs), then do not request a 2FA code for signing/sending.
Also, do not perform limits checking/calculations if 2FA is not enabled
or required.

Now that the Green backend does not require 2FA for already server-signed
txs, this allows callers to use e.g. "sign_with": ["all"] regardless of
client type, and then send the tx with send_transaction without
incurring two sets of 2FA or having limits decremented twice.

Note we still send the tx to the backend even if there is nothing to
sign. The server will validate the signatures and update internal
bookkeeping before sending, for example if we alternately sign the same
tx with different sighashes.