tx: expose scriptsig and witness computation for re-use
PSBT signing needs access to this code in order to return tx hex that
estimates fees as reliably as the create/sign/send flow now does. We
also use the "transaction" element to hold completed signatures which we
can fetch and set in the PSBT in the future.

Also convert empty sigs into dummy ones for fee estimation purposes
inside this call, so we don't have to repeat this logic in the callers.