psbt: expose a temporary endpoint for create tx json to psbt conversion
The exact interface of this call is yet to be determined, so for now we
only expose the call without documentation for experimentation by the Green
wallets and selected third parties. The interface will be finalized once
the fidelity of the conversion for the full matrix of tx features that gdk
supports is tested.

Of note: As watch-only sessions can now propose transactions (i.e use
create_transaction to create unsigned transactions), this call enables a
fully offline signing scenario whereby the WO session creates/blinds the tx
and fetches it as a PSBT to be passed to an offline full session (such as an
airgapped Jade, or offline green-cli session) to sign.