psbt: allow passing "utxos" in {"asset_id": [{utxo}, ...]} format
psbt_sign and get_details are outliers in that they currently expect the
utxos to be passed as a flattened array. Since no gdk calls return utxos
in that format, callers must munge gdk JSON to be accepted by this call.

Instead, allow the standard format so it can be passed in directly.