psbt: Make PSBT details JSON more consistent with create_tx JSON
These will never align 100% as some information in create_transaction
JSON cannot be represented (and is not relevant) when PSBT signing. For
example:

- Which outputs were greedy, if any. The output amounts are fixed by the
  time signing occurs.
- The UTXO selection criteria used, since the UTXOs to use are all given.
- The users chosen fee rate, which must be assumed to be the calculated
  fee rate.
- Change information and addressees, as these cannot be reliably
  determined (this could be improved in the future).

Nonetheless we attempt to minimise the differences so that callers can
introspect and/or display signing results from PSBTs with minimal changes
from the standard signing code path.