multisig: compute wallet identifiers earlier
Split out the code for wallet identifier calculation in order to share
it. This change makes the identifiers available earlier in the login
flow, which allows using them for e.g. blob server identifiers in the
future.

Note we now expose std::string_view literals to the gdk namespace so we
can use them as keys instead of char pointers.