blob: add and use support for merging local tx memos with blob memos
Singlesig can have multiple wallet instances each with their own
metadata. It is also (currently) possible for a wallet to switch
its blobserver setting on and off.

When we log in to a singlesig wallet which did not previously use the
client blob, merge any local tx memos with the contents of the blob,
re-save it if it was updated, and then set the local memos to the synced
collection of memos.

Also fixes a incorrectly reported race when initially populating the
client blob for the first time.