blob: delete any cached blob if we disable the blobserver for singlesig
If a user has a blobserver enabled, and then disables it, nuke any
cached client blob leaving only their local data (which was synced to
the blob when it was last enabled). When/if the user next enables the
blobserver again, any local changes will be merged with the blob on
login.