blob: expose get_accounts_settings
Singlesig currently requires the calling code to create subaccounts
before they are visible to it, despite the fact that the calling code
only knows which subaccounts to "create" because it asked the singlesig
session for the pointers in the first place.

Because of this you can't use the metadata from getsubaccounts until
they are created. So, expose get_accounts_settings to fetch just the
metadata directly.

This leads to the absurdity that you can fetch the metdata before
"creating" the subaccounts but you can't set it, since the setting code
requires the subaccount to have been "created" first. We'll work around
that in a future commit.