wamp: add is_mandatory to wamp sessions
This bool represents whether a session must be connected for the session
to function correctly. For multisig, a backend server connection is
currently mandatory. A blobserver connection is also currently mandatory
but this will be changed in a future commit.

Do not wait for mandatory connections when connecting; if the connection
fails we will handle it when we attempt to use it along with other
transient failures.