blob: rework initial blob syncing, always use a blob for singlesig
We must keep track of whether the blob needs merging, so that we can
merge it to the current blob in the event that we lose our blobserver
connection.

Also keep track of whether the blob is modified (not to be confused
with whether it is outdated with respect to any server blob).

Change singlesig to always store a local client blob, regardless of
whether we have a blobserver or not. This reduces the complexity of
merging and initially creating a blob, as well as the logic to apply
metadata to returned results.

Use the above to simplify the initial blob creation logic for singlesig,
and remove several now-uneeded c++ to rust calls.