GA_cache_control: return subaccount xpubs and origin info for singlesig
This allows us to export the subaccount names along with the xpubs. Adding
origin info is an extention to the BIP, which should be proposed in due
course. Without the origin info the path must be assumed from the
descriptor type, which makes the BIP unusable for non-BIP44-style
wallets/multisig.