rust: pass the cipher to use directly to StoreMeta
There is no need to force callers to pass an xpub here, nor to be
proscriptive on how the cipher is constructed.

Make the master xpub in the electrum session private. Its only
current use is to produce other identifiers (the cache filename and
cipher), and for DWO sessions it contains fake data. This makes sure
it can't be accessed internally which will in turn allow us to
remove/rename it later.