ga_rust: add an initial POC implementation of RWO for singlesig
As a proof of concept, this impl requires the privately derived secret
pubkey, and the blob encryption key as its credentials. The client blob
id, cache path and encryption key are derived from this pubkey in order
to load the local cache and client blob.

We create a normal "full" singlesig session, with the subaccount xpubs
taken from the client blob. At present only the first subaccount is
loaded, but this is sufficient to demonstrate logging in, fetching
addresses etc with a suitably hacked up test.

Actually loging in a singlesig RWO session is not exposed in this
commit, which is intentionally incomplete in order to keep the changes
reviewable. Further commits will extend the RWO functionality and update
the login_user interface to support RWO.