blob: unify rich watch only blobserver credential handling
RWO blobserver credentials now consist of username/password/watch_only_data,
with the latter being encrypted to the username and password so they can
be stored.

Note we maintain username and password instead of just encrypting with a
password because a password alone is unlikely to have sufficient entropy
to protect the local watch only data from offline attack.