psbt: include correct witness/non-witness utxos for inputs
non-witness UTXOs are required to mitigate the segwit fee attack and for
external signers to sign non-segwit inputs. Liquid has an explicit fee
output and is thus immune from this attack, but still requires full UTXOs
for non-segwit inputs.

Wally currently requires segwit inputs to have a witness UTXO in order
to function correctly when signing. While wally will be updated to not
require this, many other signers also have the same requirement, so we
must always provide a witness utxo for segwit inputs too.