xpub_hdkeys: set the parent fingerprint when inserting subaccounts
For multisig, we cache the subaccount/1 key, and not subaccount
directly (this is so that we only have a single derivation to produce
keys, instead of repeatedly deriving 1/pointer). We also avoid
generating fingerprints when deriving since we don't use them for the
vast majority of keys.

For recovery keys, the subaccount root xpub is the master xpub, rather
than being derived from a shared common xpub. Store the parent fingerprint
in the cached /1 key, so we can fetch it for producing recovery keypaths
when generating PSBTs,