rust: sync: improve handling of replaced txs
Suppose a wallet receives a transaction (`T`), but then the sender
replaces such transaction removing the output belonging to the
wallet. `T` must disappear from the wallet transactions list.

Before this commit we were handling this correctly if the wallet
address/script involving `T` had another transaction. But if the
wallet address/script was only involving `T` we would treat the
script as it never had any transactions and we would (incorrectly)
never remove `T`.

This commit fixes the unhandled case.