net_params: add discount_fees to determine if the Liquid fee discount is active
When enabled, Liquid fees for confidential transactions are reduced to
those of non-confidential transactions.

This is overridable by callers so can be enabled without upgrading gdk
once the feature is enabled for a given network. Once all networks are
updated, this option will be removed and always-enabled (hence, it isn't
documented as it is expected to go live shortly).