tx: fix weight calculation when all inputs are non-segwit and the tx is not yet blinded
Blinding creates output witnesses which means all inputs will be serialized
with empty witnesses, even if there are no segwit inputs. If we create a tx with
no segwit inputs (all p2sh/p2pkh inputs), we need to account for the
empty witnesses that blinding will subsequently cause to be present.

This fixes an issue where the weight of the transaction would increase
after blinding, which in turn would make the calculated fee rate lower
than the desired fee rate.