psbt_sign: finalize the PSBT after signing
Allow partial finalization if we expect the PSBT to be incomplete, as
per the previous commit.

Note that now we have finalization available via broadcast_transaction
with simulate=true, we may want to change psbt_sign to sign-only and
not finalize at all here. This is left for the future as it would be a
behaviour change callers would need to adapt to.