psbt: add signatures from signing to the PSBT instead of directly finalizing
Once all signatures (and scripts) have been added, finalize all inputs that
we have a full set of signatures for.

This allows multisig inputs that have only been user-signed to remain
un-finalized after signing, compared to the previous behaviour of
finalizing the input with a dummy (invalid) server signature. This in
turn allows server-only signing of the inputs at a later time.

PSBTs with user-signed inputs can now also be server-signed without the
user having to re-sign (i.e. re-user-sign) the input at the same time,
which is undesirable and impossible/impractical in some scenarios such
as offline signing.

This also ensures that all finalization of PSBTs is performed using the
same wally finalization logic rather than bespoke handling.

Previously, PSBT signing did not require a correctly populated PSBT in
order to sign, in that the inputs scripts could be missing. We now
populate the scripts when signing so that external finalization of
non-finalized inputs can be performed outside of gdk if desired.