build: allow skipping the source hash check for dependencies when developing
Pass '--skip-hash' to tools/builddeps.sh to skip the check.

For development iteration, it is much easier/faster to be able to modify
the source tarball of a dependency directly and rebuild, rather than
pushing the changes somewhere every time we want to test a fix.

This goes doubly for upstream repos we have to patch locally because we
don't control the sources.