taproot: support signing non-wallet inputs
Taproot signatures commit to the scriptpubkeys of all inputs in the
transaction, compared to segwit/pre-segwit which only commit to the
script of the input being signed.

So, for non-wallet or "skip_signing"=true inputs in a tx that we are
taproot signing, we need to provide the scriptpubkey.

Allow providing it in the UTXOs/tx inputs as "scriptpubkey", and
populate this for PSBT signing (where we have already ensured that
the inputs UTXO is populated). Since singlesig still doesn't cache
looked up txs, this avoids a network roundtrip for each such input.

If the caller doesn't provide it, then fetch the UTXO and grab the
scriptpubkey from it instead.