rust: remove outgoing tx signature checks
These signatures are already checked when bumping. Rather than silently
ignoring the tx, the user can refresh or change server if they ever
discover a server attempting to make them bump a fraudulent tx.

Should we choose to, we can also check the signatures when the user
fetches txs and warn if invalid sigs are found.